+++ assets/js/Containers/FrevaGPT/components/ChatComponents/UserInputBlock.js
@@ -27,12 +27,12 @@ function UserInputBlock({ content, onEdit }) {
       <>
         <FormControl
           as="textarea"
-          id={`UserInputField-${content.original_index}`}
+          id={`UserInputField-${content.user_index}`}
           className="mb-2"
           defaultValue={content.content}
           onChange={(e) => {
             handleEdit(e);
-            resizeInputField(`UserInputField-${content.original_index}`);
+            resizeInputField(`UserInputField-${content.user_index}`);
           }}
         />
 
@@ -49,7 +49,7 @@ function UserInputBlock({ content, onEdit }) {
           <Button
             variant="info"
             onClick={() => {
-              startEditedChat(editedInput, content.original_index);
+              startEditedChat(editedInput, content.user_index);
               setRenderInput(false);
             }}
           >
@@ -61,7 +61,7 @@ function UserInputBlock({ content, onEdit }) {
   }
 
   return (
-    <Col md={USER_INPUT_STYLE} key={`${content.original_index}-user`}>
+    <Col md={USER_INPUT_STYLE} key={`${content.user_index}-user`}>
       <Card className="bot-shadow br-8 card-body border-0 border-bottom bot-bg-lg">
         {renderInput
           ? renderInputComponent()