+++ assets/js/Containers/FrevaGPT/components/SidePanel/ThreadList.js
@@ -1,8 +1,6 @@
 import React from "react";
 import PropTypes from "prop-types";
 
-import { ListGroup } from "react-bootstrap";
-
 import ThreadLink from "./ThreadLink";
 
 export default function ThreadList({ threadList, setThreadList }) {
@@ -20,7 +18,7 @@ export default function ThreadList({ threadList, setThreadList }) {
   }
 
   return (
-    <ListGroup variant="flush">
+    <div>
       {threadList.map((element) => {
         return (
           <ThreadLink
@@ -30,7 +28,7 @@ export default function ThreadList({ threadList, setThreadList }) {
           />
         );
       })}
-    </ListGroup>
+    </div>
   );
 }